98 - renaming test package
