+++ trade-module/src/test/java/io/edpn/backend/trade/adapter/persistence/stationplanetaryrequest/DeleteStationRequireOdysseyRequestPortTest.java
@@ -1,8 +1,8 @@
-package io.edpn.backend.trade.adapter.persistence.stationrequireodysseyrequest;
+package io.edpn.backend.trade.adapter.persistence.stationplanetaryrequest;
 
 import io.edpn.backend.trade.adapter.persistence.StationPlanetaryRequestRepository;
 import io.edpn.backend.trade.adapter.persistence.repository.MybatisStationPlanetaryRequestRepository;
-import io.edpn.backend.trade.application.port.outgoing.stationplanetaryrequest.CreateStationPlanetaryRequestPort;
+import io.edpn.backend.trade.application.port.outgoing.stationplanetaryrequest.DeleteStationRequireOdysseyRequestPort;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
@@ -12,12 +12,12 @@ import org.mockito.junit.jupiter.MockitoExtension;
 import static org.mockito.Mockito.verify;
 
 @ExtendWith(MockitoExtension.class)
-public class CreateStationPlanetaryRequestPortTest {
+public class DeleteStationRequireOdysseyRequestPortTest {
 
     @Mock
     private MybatisStationPlanetaryRequestRepository mybatisStationPlanetaryRequestRepository;
 
-    private CreateStationPlanetaryRequestPort underTest;
+    private DeleteStationRequireOdysseyRequestPort underTest;
 
     @BeforeEach
     public void setup() {
@@ -29,9 +29,9 @@ public class CreateStationPlanetaryRequestPortTest {
         String systemName = "someName";
         String stationName = "someName2";
 
-        underTest.create(systemName, stationName);
+        underTest.delete(systemName, stationName);
 
-        verify(mybatisStationPlanetaryRequestRepository).insert(systemName, stationName);
+        verify(mybatisStationPlanetaryRequestRepository).delete(systemName, stationName);
     }
